LexMake: Fix unstyled first Line VCBuild: Compile fix

+++ src.3.8.0.lua53/src/scintilla/lua/src/lpvm.c
@@ -67,7 +67,7 @@ static Stack *doublestack (lua_State *L, Stack **stacklimit, int ptop) {
   int n = *stacklimit - stack;  /* current stack size */
   int max, newn;
   lua_getfield(L, LUA_REGISTRYINDEX, MAXSTACKIDX);
-  max = lua_tointeger(L, -1);  /* maximum allowed size */
+  max = (int)lua_tointeger(L, -1);  /* maximum allowed size */
   lua_pop(L, 1);
   if (n >= max)  /* already at maximum size? */
     luaL_error(L, "too many pending calls/choices");
@@ -102,7 +102,7 @@ static int resdyncaptures (lua_State *L, int fr, int curr, int limit) {
       luaL_error(L, "invalid position returned by match-time capture");
   }
   lua_remove(L, fr);  /* remove first result (offset) */
-  return res;
+  return (int)res;
 }
